Separate out invalidateCacheGroup from main data.
