+++ assets/js/modules/analytics/datastore/settings.js
@@ -21,7 +21,8 @@
  */
 import API from 'googlesitekit-api';
 import Data from 'googlesitekit-data';
-import dataAPI, { TYPE_MODULES } from '../../../components/data';
+import { TYPE_MODULES } from '../../../components/data/constants';
+import { invalidateCacheGroup } from '../../../components/data/invalidate-cache-group';
 import {
 	isValidAccountID,
 	isValidInternalWebPropertyID,
@@ -110,7 +111,7 @@ export const controls = {
 
 		await API.invalidateCache( 'modules', 'analytics' );
 		// TODO: Remove once legacy dataAPI is no longer used.
-		dataAPI.invalidateCacheGroup( TYPE_MODULES, 'analytics' );
+		invalidateCacheGroup( TYPE_MODULES, 'analytics' );
 
 		return {};
 	} ),
